refactor: extract FrameAdjustModal, FrameScheduleModal, FramesModal components
